requirements.txt
google-generativeai
pydantic
python-dotenv
firebase-admin